HAPP-1332: Create FetchSpecialAuthoritiesUseCase
